[ledger] test with PersistentStorage
