+++ apps/src/bin/anoma-node/shell/storage/types.rs
@@ -145,28 +145,6 @@ impl<I> PrefixIterator<I> {
     }
 }
 
-impl<'a> Iterator for PrefixIterator<rocksdb::DBIterator<'a>> {
-    type Item = (String, Vec<u8>, u64);
-
-    /// Returns the next pair and the gas cost
-    fn next(&mut self) -> Option<(String, Vec<u8>, u64)> {
-        match self.iter.next() {
-            Some((key, val)) => {
-                let key = String::from_utf8(key.to_vec())
-                    .expect("Cannot convert from bytes to key string");
-                match key.strip_prefix(&self.db_prefix) {
-                    Some(k) => {
-                        let gas = k.len() + val.len();
-                        Some((k.to_owned(), val.to_vec(), gas as _))
-                    }
-                    None => self.next(),
-                }
-            }
-            None => None,
-        }
-    }
-}
-
 impl<I> std::fmt::Debug for PrefixIterator<I> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_str("PrefixIterator")
